Rework canvas access warning
The previous implementation was checking whether a setting like
resistfingerprinting was preventing access to the browser's performance
API.  However, it's possible to have resistfingerprinting enabled but
still enable canvas access on a site-by-site basis (e.g. the browser
prompts for permission upon first access if resistfingerprinting is
enabled).

This commit changes the access check to directly examine the result of
trying to extract data from a fake, blank canvas. If anything but zeroes
appear when checking data, such as the random integers Firefox returns,
then it can safely determined that the browser is blocking canvas
access.

For the wordcloud homepage, this means that because a timer is employed,
enabling the canvas access immediately sees the wordcloud starting to
work, without even needing to reload.

Ref: #44